Add `id`'s and `href` to HTML headings
This patch adds `id` tags to headings, and make the content of the
heading a link to itself, i.e. `<hx>Title</hx>` is replaced by
`<hx id="Title"><a href="#Title">Title</a></hx>`. This make it possible
to refer directly to sub sections (e.g.
https://defelement.com/ciarlet.html#Mapping+finite+elements will link
directly to the "Maping finite elements" section). The CSS is tweaked so
that the newly added links do not cause headings to be underlined.